# tracer: support dot notation for tags with array values