--- a/ddtrace/tracer/span.go
+++ b/ddtrace/tracer/span.go
@@ -141,22 +141,6 @@ func (s *span) SetTag(key string, value interface{}) {
 		s.setMeta(key, v)
 		return
 	}
-	if value != nil {
-		switch reflect.TypeOf(value).Kind() {
-		case reflect.Slice:
-			slice := reflect.ValueOf(value)
-			for i := 0; i < slice.Len(); i++ {
-				key := fmt.Sprintf("%s.%d", key, i)
-				v := slice.Index(i)
-				if num, ok := toFloat64(v.Interface()); ok {
-					s.setMetric(key, num)
-				} else {
-					s.setMeta(key, fmt.Sprintf("%s", v))
-				}
-			}
-			return
-		}
-	}
 	if v, ok := toFloat64(value); ok {
 		s.setMetric(key, v)
 		return
@@ -176,6 +160,25 @@ func (s *span) SetTag(key string, value interface{}) {
 		}()
 		s.setMeta(key, v.String())
 		return
+	}
+	if value != nil {
+		// Arrays will be translated to dot notation. e.g.
+		// { "myarr.0": "foo", "myarr.1": "bar"}
+		// which will be displayed as an array in the UI.
+		switch reflect.TypeOf(value).Kind() {
+		case reflect.Slice:
+			slice := reflect.ValueOf(value)
+			for i := 0; i < slice.Len(); i++ {
+				key := fmt.Sprintf("%s.%d", key, i)
+				v := slice.Index(i)
+				if num, ok := toFloat64(v.Interface()); ok {
+					s.setMetric(key, num)
+				} else {
+					s.setMeta(key, fmt.Sprintf("%s", v))
+				}
+			}
+			return
+		}
 	}
 	// not numeric, not a string, not a fmt.Stringer, not a bool, and not an error
 	s.setMeta(key, fmt.Sprint(value))