--- a/ddtrace/tracer/span_test.go
+++ b/ddtrace/tracer/span_test.go
@@ -337,6 +337,21 @@ func TestSpanSetTag(t *testing.T) {
 	span.SetTag("nilStringer", (*nilStringer)(nil))
 	assert.Equal("<nil>", span.Meta["nilStringer"])
 
+	span.SetTag("somestrings", []string{"foo", "bar"})
+	assert.Equal("foo", span.Meta["somestrings.0"])
+	assert.Equal("bar", span.Meta["somestrings.1"])
+
+	span.SetTag("somenums", []int{-1, 5, 2})
+	assert.Equal(-1., span.Metrics["somenums.0"])
+	assert.Equal(5., span.Metrics["somenums.1"])
+	assert.Equal(2., span.Metrics["somenums.2"])
+
+	span.SetTag("someslices", [][]string{{"a, b, c"}, {"d"}, nil, {"e, f"}})
+	assert.Equal("[a, b, c]", span.Meta["someslices.0"])
+	assert.Equal("[d]", span.Meta["someslices.1"])
+	assert.Equal("[]", span.Meta["someslices.2"])
+	assert.Equal("[e, f]", span.Meta["someslices.3"])
+
 	assert.Panics(func() {
 		span.SetTag("panicStringer", &panicStringer{})
 	})