<!--
* New contributors are highly encouraged to read our
  [CONTRIBUTING](/CONTRIBUTING.md) documentation.
* Commit and PR titles should be prefixed with the general area of the pull request's change.

-->
### What does this PR do?

<!--
* A brief description of the change being made with this pull request.
* If the description here cannot be expressed in a succinct form, consider
  opening multiple pull requests instead of a single one.
-->

Improves array support in tag values by automatically encoding arrays with dot notation.

With the following code snippet:
```
span := tracer.StartSpan("somename", tracer.Tag("custom", []string{"foo", "bar"}))
```

Before (just a string):
<img width="193" alt="Screenshot 2023-10-06 at 2 53 53 PM" src="https://github.com/DataDog/dd-trace-go/assets/28789760/94e6b298-ce94-4097-9eaf-0c0bd53b7104">

After (an array):
<img width="193" alt="Screenshot 2023-10-06 at 2 53 45 PM" src="https://github.com/DataDog/dd-trace-go/assets/28789760/bfbfc3bc-8d14-4531-983f-7c97f3011f2d">

Note that this only supports one level of array nesting. Any more, and the user will need to do this manually.

AIT-8525

### Motivation

<!--
* What inspired you to submit this pull request?
* Link any related GitHub issues or PRs here.
* If this resolves a GitHub issue, include "Fixes #XXXX" to link the issue and auto-close it on merge.
-->

This supports more precise search queries and filtering. For example, with the previous code, a `DD_APM_REPLACE_TAG` with `custom` would have removed the entire array. Now,  a `DD_APM_REPLACE_TAG` with `custom` will do nothing. A customer will instead specify the index, e.g. a `DD_APM_REPLACE_TAG` with `custom.1`, to only redact the indices they want, or do `custom.*` to redact them all.

### Reviewer's Checklist
<!--
* Authors can use this list as a reference to ensure that there are no problems
  during the review but the signing off is to be done by the reviewer(s).
-->

- [ ] Changed code has unit tests for its functionality at or near 100% coverage.
- [ ] There is a benchmark for any new code, or changes to existing code.
- [ ] If this interacts with the agent in a new way, a system test has been added.

For Datadog employees:

- [ ] If this PR touches code that handles credentials of any kind, such as Datadog API keys, I've requested a review from `@DataDog/security-design-and-guidance`.
- [ ] This PR doesn't touch any of that.

Unsure? Have a question? Request a review!