@katiehockman 2023-11-14
Because these changes will cause `reflect` to be called on every tag, there is a small performance cost if the code is reached. For example, if the code is above the `Stringer` check, then it adds a check every time a tag is added to see if it's a slice.
```
benchstat old.txt new.txt
goos: darwin
goarch: arm64
pkg: gopkg.in/DataDog/dd-trace-go.v1/ddtrace/tracer
                  │   old.txt   │              new.txt               │
                  │   sec/op    │   sec/op     vs base               │
SetTagMetric-10     42.37n ± 2%   42.74n ± 3%       ~ (p=0.631 n=10)
SetTagString-10     39.99n ± 3%   39.81n ± 3%       ~ (p=0.579 n=10)
SetTagStringer-10   54.71n ± 2%   56.82n ± 2%  +3.87% (p=0.000 n=10)
SetTagField-10      18.61n ± 1%   18.55n ± 2%       ~ (p=0.361 n=10)
geomean             36.24n        36.60n       +0.97%
```

I'm not terribly concerned by 2 nanoseconds, but it could in theory add up if someone has many tags. I moved the check to be as low down the list of checks as possible so it's hopefully not hit in most cases (I expect the majority of tags are strings, ints, floats).