@katiehockman 2023-11-14
The parametric tests caught a bug in the way we were encoding the value, which should now be fixed. The system-tests PR is here: https://github.com/DataDog/system-tests/pull/1841