@katiehockman 2023-11-14
And the failing parametric test is because that relied on old behavior. https://github.com/DataDog/system-tests/pull/1841 will skip that test and enable the new one.